deleted files: `src/app/gui/wms/service.js` and  `src/app/gui/wms/vue/panel/wmslayerspanel.js`
replaced by: `src/app/gui/wms/vue/wms.js`
